LayoutTests/inspector-protocol/dom/getAccessibilityPropertiesForNode_liveRegion-expected.txt

Checking Web Inspector protocol (specifically live region properties) for the Accessibility Node Inspector.

Total elements to be tested: 17.

<div role="timer">off (default)</div>
    exists: true

<div role="status" aria-live="off">off</div>
    exists: true

<div role="status" aria-live="assertive">assertive</div>
    exists: true
    liveRegionAtomic: true
    liveRegionStatus: assertive

<div role="status">polite (default)</div>
    exists: true
    liveRegionAtomic: true
    liveRegionStatus: polite

<div role="marquee">off (default)</div>
    exists: true

<div role="log" aria-live="off">off</div>
    exists: true

<div role="log" aria-live="assertive">assertive</div>
    exists: true
    liveRegionAtomic: false
    liveRegionStatus: assertive

<div role="log">polite (default)</div>
    exists: true
    liveRegionAtomic: false
    liveRegionStatus: polite

<div role="alert" aria-live="polite">polite</div>
    exists: true
    liveRegionAtomic: true
    liveRegionStatus: polite

<div role="alert" aria-live="off">off</div>
    exists: true

<div role="alert">assertive (default)</div>
    exists: true
    liveRegionAtomic: true
    liveRegionStatus: assertive

<div role="group" aria-live="assertive" aria-busy="true" aria-atomic="false">assertive</div>
    exists: true
    busy: true
    liveRegionAtomic: false
    liveRegionStatus: assertive

<div role="group" aria-live="polite" aria-busy="true" aria-atomic="false">polite</div>
    exists: true
    busy: true
    liveRegionAtomic: false
    liveRegionStatus: polite

<div role="group" aria-live="off" aria-busy="true" aria-atomic="false">off</div>
    exists: true
    busy: true

<div role="group" aria-live="assertive" aria-busy="true" aria-atomic="true">assertive</div>
    exists: true
    busy: true
    liveRegionAtomic: true
    liveRegionStatus: assertive

<div role="group" aria-live="polite" aria-busy="true" aria-atomic="true">polite</div>
    exists: true
    busy: true
    liveRegionAtomic: true
    liveRegionStatus: polite

<div role="group" aria-live="off" aria-busy="true" aria-atomic="true">off</div>
    exists: true
    busy: true
